@@ -23,14 +23,14 @@
     "@types/lodash": "^4.14.172",
     "@types/node": "^18.6.4",
     "@types/qs": "^6.9.7",
-    "@types/react": "^17.0.0",
+    "@types/react": "^18.2.15",
     "@types/react-copy-to-clipboard": "^5.0.2",
-    "@types/react-dom": "^17.0.0",
+    "@types/react-dom": "^18.2.7",
     "@types/react-grid-layout": "^1.3.2",
     "@types/react-lazylog": "^4.5.1",
     "@types/react-plotly.js": "^2.5.0",
-    "@types/react-router": "^5.1.18",
-    "@types/react-router-dom": "^5.1.8",
+    "@types/react-router": "^5.1.20",
+    "@types/react-router-dom": "^5.3.3",
     "@types/react-syntax-highlighter": "^13.5.2",
     "@types/react-virtualized-auto-sizer": "^1.0.1",
     "@types/react-window": "^1.8.5",
@@ -74,18 +74,18 @@
     "rc-field-form": "^1.27.1",
     "rc-image": "^5.2.5",
     "rc-steps": "^6.0.0",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-code-blocks": "^0.0.9-0",
     "react-color": "^2.19.3",
     "react-copy-code": "^2.1.2",
     "react-copy-to-clipboard": "^5.0.4",
     "react-dnd": "^16.0.1",
     "react-dnd-html5-backend": "^16.0.1",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-dropzone": "^14.2.3",
     "react-grid-layout": "^1.3.4",
     "react-hooks-global-state": "^1.0.1",
-    "react-i18next": "^11.11.4",
+    "react-i18next": "^13.0.2",
     "react-icons": "^4.2.0",
     "react-in-viewport": "^1.0.0-alpha.28",
     "react-is": ">= 16.8.0",
@@ -99,7 +99,7 @@
     "react-refresh": "^0.14.0",
     "react-resizable": "^3.0.4",
     "react-router": "5.3.4",
-    "react-router-dom": "^5.2.0",
+    "react-router-dom": "^5.3.4",
     "react-sortablejs": "^6.1.4",
     "react-syntax-highlighter": "^15.4.5",
     "react-uid": "^2.3.2",
