@@ -39,6 +39,8 @@
   "dependencies": {
     "@rjsf/utils": "^5.0.0-beta.13",
     "@sentry/browser": "6.19.7",
+    "@types/react": "^18.2.15",
+    "@types/react-dom": "^18.2.7",
     "baseui": "^12.2.0",
     "fast-deep-equal": "^3.1.3",
     "history": "4.10.1",
@@ -46,6 +48,8 @@
     "klona": "^2.0.5",
     "lodash": "4.17.21",
     "nanoid": "^4.0.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-in-viewport": "^1.0.0-alpha.28",
     "react-resizable": "^3.0.4",
     "react-sortablejs": "^6.1.4",
@@ -64,15 +68,11 @@
     "@types/history": "4.7.11",
     "@types/jest": "28.1.6",
     "@types/lodash": "4.14.182",
-    "@types/react": "17.0.42",
-    "@types/react-dom": "17.0.14",
     "@types/sortablejs": "^1.15.0",
     "@types/systemjs": "^0.20.6",
     "@vitejs/plugin-react": "^2.2.0",
     "esbuild": "0.15.7",
     "lodash": "4.17.21",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
     "rimraf": "3.0.2",
     "rollup": "2.79.1",
     "rollup-plugin-dts": "^4.2.2",
@@ -83,9 +83,5 @@
     "typescript": "^4.9.3",
     "vite": "^3.2.2",
     "vite-plugin-dts": "^1.7.1"
-  },
-  "peerDependencies": {
-    "react": "17.0.2",
-    "react-dom": "17.0.2"
   }
 }
