@@ -140,7 +140,12 @@ const hijacked = {
     invalidFile,
 }
 
-export default function IconFont({ size = 14, type = 'user', kind = 'inherit', style = {} }: IIconFontProps) {
+export default function IconFont({
+    size = 14,
+    type = 'user',
+    kind = 'inherit',
+    style = {},
+}: IIconFontProps): React.ReactNode | any {
     const [, theme] = themedUseStyletron()
 
     const colors = {
