@@ -5,6 +5,7 @@ import { WidgetConfig, WidgetGroupType, WidgetRendererProps } from '@starwhale/c
 import { WidgetPlugin } from '@starwhale/core/widget'
 import { UI_DATA } from '@starwhale/core/form/schemas/fields'
 import { getRocAucConfig } from '@starwhale/ui/Plotly/utils'
+import { useDatastoreDecodeRecords } from '@starwhale/core/datastore'
 
 const PlotlyViewer = React.lazy(() => import(/* webpackChunkName: "PlotlyViewer" */ '@starwhale/ui/Plotly'))
 
@@ -57,7 +58,9 @@ function PanelRocAucWidget(props: WidgetRendererProps<any, any>) {
     const { data: formData } = fieldConfig ?? {}
     const { chartTitle: title, x, y } = formData ?? {}
 
-    const rocAucData = useParseRocAuc({ records, x, y })
+    const $records = useDatastoreDecodeRecords(records)
+
+    const rocAucData = useParseRocAuc({ records: $records, x, y })
     const vizData = getRocAucConfig(title, [], rocAucData as any)
 
     return (
